Fix stats script
